Create addresses module for CanonicalAddr, HumanAddr
