+++ packages/std/src/imports.rs
@@ -1,5 +1,6 @@
 use std::vec::Vec;
 
+use crate::addresses::{CanonicalAddr, HumanAddr};
 use crate::encoding::Binary;
 use crate::errors::{StdError, StdResult};
 #[cfg(feature = "iterator")]
@@ -7,7 +8,6 @@ use crate::iterator::{Order, KV};
 use crate::memory::{alloc, build_region, consume_region, Region};
 use crate::serde::from_slice;
 use crate::traits::{Api, Querier, QuerierResult, ReadonlyStorage, Storage};
-use crate::types::{CanonicalAddr, HumanAddr};
 
 /// An upper bound for typical canonical address lengths (e.g. 20 in Cosmos SDK/Ethereum or 32 in Nano/Substrate)
 const CANONICAL_ADDRESS_BUFFER_LENGTH: usize = 32;
